# control-api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.createRequestEventRequest:
    properties:
      event_type:
        type: string
      metadata:
        type: object
      request_log_id:
        type: string
    required:
    - event_type
    - request_log_id
    type: object
  handlers.createRequestLogRequest:
    properties:
      amount_usd:
        type: string
      client_ip:
        type: string
      error_message:
        type: string
      error_type:
        type: string
      final_status_code:
        type: integer
      id:
        type: string
      method:
        type: string
      outbound_route_id:
        type: string
      path:
        type: string
      payment_channel_asset_id:
        type: string
      payment_channel_id:
        type: string
      payment_completed:
        type: boolean
      payment_completed_at:
        type: string
      payment_requested_at:
        type: string
      payment_required:
        type: boolean
      project_id:
        type: string
      request_id:
        type: string
      status:
        type: string
      upstream_response_time_ms:
        type: integer
      upstream_status_code:
        type: integer
      upstream_url:
        type: string
      user_agent:
        type: string
    required:
    - method
    - path
    - project_id
    - request_id
    - status
    type: object
...
  handlers.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.requestEventResponse:
    properties:
      created_at:
        type: string
      event_type:
        type: string
      id:
        type: string
      metadata:
        type: object
      request_log_id:
        type: string
    type: object
  handlers.requestLogResponse:
    properties:
      amount_usd:
        type: string
      client_ip:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      error_type:
        type: string
      final_status_code:
        type: integer
      id:
        type: string
      method:
        type: string
      outbound_route_id:
        type: string
      path:
        type: string
      payment_channel_asset_id:
        type: string
      payment_channel_id:
        type: string
      payment_completed:
        type: boolean
      payment_completed_at:
        type: string
      payment_requested_at:
        type: string
      payment_required:
        type: boolean
      project_id:
        type: string
      request_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      upstream_response_time_ms:
        type: integer
      upstream_status_code:
        type: integer
      upstream_url:
        type: string
      user_agent:
        type: string
    type: object
...
  handlers.updateRequestLogRequest:
    properties:
      amount_usd:
        type: string
      error_message:
        type: string
      error_type:
        type: string
      final_status_code:
        type: integer
      outbound_route_id:
        type: string
      payment_channel_asset_id:
        type: string
      payment_channel_id:
        type: string
      payment_completed:
        type: boolean
      payment_completed_at:
        type: string
      payment_requested_at:
        type: string
      payment_required:
        type: boolean
      status:
        type: string
      upstream_response_time_ms:
        type: integer
      upstream_status_code:
        type: integer
      upstream_url:
        type: string
    required:
    - status
    type: object
...
host: localhost:9090
info:
  contact: {}
  description: Control plane for xpaywall — manages projects, routes, users, payment
    channels, and request logs.
  title: xpaywall Control API
  version: "1.0"
paths:
...
  /api/v1/request-events:
    post:
      consumes:
      - application/json
      parameters:
      - description: Event data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createRequestEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.requestEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create request event
      tags:
      - request-logs
  /api/v1/request-logs:
    get:
      parameters:
      - description: Max results (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by Project ID (UUID)
        in: query
        name: project_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.requestLogResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List request logs
      tags:
      - request-logs
    post:
      consumes:
      - application/json
      parameters:
      - description: Request log data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createRequestLogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.requestLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create request log
      tags:
      - request-logs
  /api/v1/request-logs/{id}:
    get:
      parameters:
      - description: Request Log ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.requestLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get request log
      tags:
      - request-logs
    put:
      consumes:
      - application/json
      parameters:
      - description: Request Log ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateRequestLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.requestLogResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update request log
      tags:
      - request-logs
  /api/v1/request-logs/{id}/events:
    get:
      parameters:
      - description: Request Log ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.requestEventResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List request events
      tags:
      - request-logs
...
securityDefinitions:
  ApiKeyAuth:
    description: Internal API key shared between xgateway and control-api
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: 'JWT token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"